requirements.txt
fastapi
uvicorn
jinja2
python-multipart
requests
pymongo
cloudinary
python-dotenv